Replace hash approach

+++ bundles/sirix-core/src/main/java/org/sirix/access/trx/node/ForwardingNodeReadOnlyTrx.java
@@ -123,7 +123,7 @@ public interface ForwardingNodeReadOnlyTrx extends NodeReadOnlyTrx {
     }
 
     @Override
-    default BigInteger getHash() {
+    default long getHash() {
         return delegate().getHash();
     }
