+++ bundles/sirix-core/src/main/java/org/sirix/index/path/summary/AbstractForwardingPathSummaryReader.java
@@ -60,7 +60,7 @@ public abstract class AbstractForwardingPathSummaryReader extends ForwardingObje
   }
 
   @Override
-  public BigInteger getHash() {
+  public long getHash() {
     return delegate().getHash();
   }
