+++ bundles/sirix-rest-api/src/main/kotlin/org/sirix/rest/crud/xml/XmlUpdate.kt
@@ -121,7 +121,7 @@ class XmlUpdate(private val location: Path) {
                             val hashCode = ctx.request().getHeader(HttpHeaders.ETAG)
                                 ?: throw IllegalStateException("Hash code is missing in ETag HTTP-Header.")
 
-                            if (wtx.hash != BigInteger(hashCode)) {
+                            if (wtx.hash != hashCode.toLong()) {
                                 throw IllegalArgumentException("Someone might have changed the resource in the meantime.")
                             }
                         }
